Assign scRNA-seq to clone-of-origin

clonealign assigns single cells (measured with RNA-seq) to their clones of origin, where the clones have been inferred from ultra-shallow scDNA-seq and collated into copy number profiles.

clonealign(gene_expression_data, copy_number_data, max_iter_em = 100,
  max_iter_adam = 1, rel_tol_em = 1e-06, rel_tol_adam = 1e-06,
  gene_filter_threshold = 0, learning_rate = 0.1, fix_alpha = TRUE,
  clone_specific_phi = TRUE, fix_s = NULL, sigma_hyper = 1,
  dtype = "float64", saturate = TRUE, verbose = TRUE)

Arguments

gene_expression_data
A matrix of gene counts or a SingleCellExperiment. This should contain raw counts. See details.

copy_number_data
A matrix or data frame of copy number calls for each clone. See details.

max_iter_em
Maximum number of EM iterations to perform

max_iter_adam
The maximum number of Adam iterations to perform in each M-step

rel_tol_em
Relative tolerance (change in log-likelihood per EM iteration in percent) below which the EM algorithm is considered converged

rel_tol_adam
The relative tolerance for each Adam update in the M-step below which the maximization will be considered converged

gene_filter_threshold
Genes with total counts below or equal to this threshold will be filtered out (removes genes with no counts by default)

learning_rate
The learning rate to be passed to the Adam optimizer

fix_alpha
Should the underlying priors for clone frequencies be fixed? Default TRUE (values are inferred from the data)

clone_specific_phi
Should each clone have distinct over-dispersion parameters? Default TRUE

fix_s
Should the size factors be fixed? If NULL they are jointly inferred from the data, otherwise a vector corresponding to the number of cells should be specified.

sigma_hyper
The hyperparameter governing shrinkage of clone-specific dispersions

dtype
The dtype for tensorflow useage, either "float32" or "float64"

saturate
Should the CNV-expression relationship saturate above copy number = 4? Default TRUE

verbose
Should warnings and EM convergence information be printed? Default TRUE

Value

An object of class clonealign_fit. The maximum likelihood estimates of the clone assignment paramters are in the clone slot. Maximum likelihood estimates of all model parameters are in the ml_params slot.

Details

Input format

gene_expression_data must either be a SingleCellExperiment or SummarizedExperiment with a counts assay representing raw gene expression counts, or a cell by gene matrix of raw counts.

copy_number_data must either be a matrix, data.frame or DataFrame with a row for each gene in gene_expression_data and a column for each of the clones. If colnames(copy_number_data) is not NULL then these names will be used for each of the clones in the final output.

Recommended parameter settings

As with any probabilistic model there are many parameters to set. Through comprehensive simulations regarding the robustness of the model to mis-specification (ie what's the minimum proportion of genes for which the CNV-expression relationship can be true and our inferences still valid) we have come up with the following guidelines for parameter settings, reflected in the default values:

Number of ADAM iterations - if set to 1 we essentially perform gradient descent on the marginal log-likelihood which empircally appears to have the best performance

Dispersions should be clone-specific with weak shrinkage (sigma = 1 appears best)

The generating probabilities should be fixed to be a priori equal (this corresponds to setting alpha = TRUE)

The cell size factors are best fixed in advanced by multiplying the total counts of whatever genes are passed to clonealign by the edgeR (TMM) normalization factors

Controlling the EM algorithm

Inference is performed using the EM algorithm (https://en.wikipedia.org/wiki/Expectation-maximization_algorithm) which uses the log marginal likelihood to monitor convergence. This is controlled using the rel_tol_em parameter. When the difference between the new and old log marginal likelihoods normalized by the absolute value of the old falls below rel_tol_em, the EM algorithm is considered converged. The maximum number of iterations to acheive this is set using the max_iter_em parameter.

In each M-step, maximization is performed using Adam, with learning rate given by learning_rate. Each M-step is considered converged with the value of the Q function falls below rel_tol_adam with a maximum number of iterations max_iter_adam.

Examples

library(SingleCellExperiment)
data(example_sce)
copy_number_data <- rowData(example_sce)[,c("A", "B", "C")]
cal <- clonealign(example_sce, copy_number_data)

#> Removing 0 genes with low counts

#> Creating Tensorflow graph...

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> Warning: Maximim number of ADAM iterations reached reached

#> clonealign inference complete

print(cal)

#> A clonealign_fit for 200 cells, 100 genes, and 3 clones
#> To access clone assignments, call x$clone
#> To access ML parameter estimates, call x$ml_params

clones <- cal$clone